+++ Monofoxe/Monofoxe.Engine.WindowsDX/Implementations/WindowsTextInputBinder.cs
@@ -1,0 +1,10 @@
+﻿namespace Monofoxe.Engine.WindowsDX.Implementations
+{
+	public class WindowsTextInputBinder : ITextInputBinder
+	{
+		public void Init()
+		{
+			GameMgr.Game.Window.TextInput += Input.TextInput;
+		}
+	}
+}