+++ Monofoxe/Monofoxe.Engine/Abstractions/IMonitorFrequencyManager.cs
@@ -1,0 +1,11 @@
+﻿
+namespace Monofoxe.Engine.Abstractions
+{
+	/// <summary>
+	/// Interface for accessing the current monitor's refresh rate.
+	/// </summary>
+	public interface IMonitorFrequencyManager
+	{
+		int GetCurrentMonitorFrequency();
+	}
+}