+++ Monofoxe/Monofoxe.Engine/EntityExtensions.cs
@@ -1,0 +1,11 @@
+﻿using Monofoxe.Engine.EC;
+using Monofoxe.Engine.Utils;
+
+namespace Monofoxe.Engine
+{
+	public static class EntityExtensions
+	{
+		public static PositionComponent GetPosition(this Entity entity) =>
+			entity.GetComponent<PositionComponent>();
+	}
+}