+++ Samples/Monofoxe.Samples.WindowsGL/Program.cs
@@ -1,0 +1,18 @@
+﻿using Monofoxe.Engine.WindowsGL;
+using System;
+
+namespace Monofoxe.Samples.WindowsGL
+{
+	public static class Program
+	{
+		[STAThread]
+		static void Main()
+		{
+			MonofoxePlatform.Init();
+			using (var game = new Game1())
+			{
+				game.Run();
+			}
+		}
+	}
+}